fragmented remote cmds
